Add fragment and bottomnavigationview
