+++ app/src/main/res/drawable-anydpi/ic_saved_words.xml
@@ -1,0 +1,14 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="28.8"
+    android:viewportHeight="28.8"
+    android:tint="#0D26EE"
+    android:alpha="0.8">
+  <group android:translateX="2.4"
+      android:translateY="2.4">
+      <path
+          android:fillColor="#FF000000"
+          android:pathData="M17,3H7c-1.1,0 -1.99,0.9 -1.99,2L5,21l7,-3 7,3V5c0,-1.1 -0.9,-2 -2,-2z"/>
+  </group>
+</vector>
